Refactor from monolith to decoupled components per file
